code/agent/requirements.txt
# Minimal dependencies inferred
python-dotenv
mistralai
requests
pydantic
fastapi

# JWT verification
cryptography
PyJWT

# Optional for ASGI support
uvicorn[standard]
